refactor:move handleSetImages into drag-drop component
